+++ source/GLaDOS.cpp
@@ -1,4 +1,4 @@
-﻿#include "GMController.h"
+﻿#include "GLaDOS.h"
 
 #include <array>
